Add likely macro, force-no-inline on some sim stuff
Makes compilation a lot faster, and also makes runtime a lot faster on
my computer. Probably due to fragile behavior of inlining on the big
switch statements, but still worth doing. -Os without -flto on clang is
a slightly smaller binary that's just as fast, but I don't know how
fragile that is to change and just happens to be good today. Need to get
a workflow for going through and looking at the output asm on Linux.