Change --seed to use atoi, change init_seed to int
This avoids using `long` in orca-c, which had previous been avoided.
Using `long` comes with some baggage. It would be best to avoid it where
possible, especially since there are no other uses of it orca-c. If
explicit sizes are needed, use the types from stdint.h. The only reason
`int` is used for argument handling in orca-c's TUI is that `atoi`
requires it. `atol` doesn't provide any advantage over `atoi`, except to
allow the `--seed` option to be larger on some platforms and some
configurations. But I think it's better to have consistent argument
handling, and have `--seed` always have a range of `0..INT_MAX`, which
is going to be signed 32-bit max on almost every platform, unlike
`long`.

This change also allows init_seed to be set to 0, where it previously
was not allowed. The default still remains 1.