Reorder `random_seed` field to waste less space.
The `Ged` struct's `random_seed` field was mistakenly placed after the
bitfield flags when it was added. It should not have been -- this wastes
space. This change moves it alongside other same-sized fields.

This change also cleans up some initialization ordering that had fallen
out of sync with the declaration ordering. This was harmless, but made
it a little bit more confusing to read.