Fix potential MIDI notes stuck on due to PortMidi latency
Since we changed to use PortMidi's timestamping and buffering stuff,
(because otherwise MIDI timing gets messed up for some people for some
unknown reason), there's now a chance that MIDI notes can get stuck on
when quitting orca, because PortMidi seems to be dumb and won't
immediately send the pending events. So we have to manually spin and
wait for them to be sent. There doesn't appear to be a way to forcefully
flush them, so we're stuck adding an delay.