Fix unused function warning with --no-mouse
ged_mouse_event() is never called when building with --no-mouse or when
defining FEAT_NOMOUSE. This could cause compilers to issue a warning.
This commit adds the ORCA_OK_IF_UNUSED decoration to ged_mouse_event(),
which should suppress the warning on most compilers. This intentionally
leaves the code compiled in. This is to avoid interleaving more ifdefs
than necessary into the code and making development and compile-time
errors more difficult to deal with.