Change TUI menu code to not use 'menu' library
This commit implements the Qnav/Qmenu system directly in ncurses,
replacing the old implementation which relied on the 'menu' library.
The new code is shorter, easier to read, and doesn't need the 'menu'
library to be linked.

For users, the behavior of the menus should be the same.

We still rely on the 'form' library for text field input in the menus.