Fix minimum-size attribute (currently unused) in gcc
This attribute macro, ORCA_OPT_MINSIZE, isn't currently in use in any of
the code, though it occasionally gets used when testing things. I had
previously only tested this under clang, and hadn't tested it under gcc.
Some of the syntax was missing for gcc's version. This commit fixes it.